Refactor internal post logic
SourceDir, and SiteDir are now package level variables for the
directories where the posts will be stored, both unpublished, and
published respectively.

filepath, and os.PathSeparator are now used for the creation of file
paths.

The post.NewFromPath function now just takes the path to the source
Markdown file for that post. The post.New function now just takes the
title, and the category strings.