Refactor util package
Added the Exit function which will take a message and an error. If the
given error is nil then the program will exit with the given message,
and an exit code of 0, otherwise it will exit with a non-zero exit code.

Add MustBeInitialized to the util package, along with OpenInEditor, and
RenderTemplate.

The dependency on go-slugify has been removed in favour of a Slug
function, since it's a fairly simple thing to do anyway, and doesn't
warrant a library.