Refactor entire code base, remove unit tests
Removed unit tests, I intend on replacing these with functionality
tests. These will most likely be a collection of shell scripts that will build,
and then run the program.

The main config file is now less bloated. Index layouts are managed at
the filesystem level, instead of being all shoved into the main config
file. So when defining a layout for an index of posts simply place your
layouts inside an '_index' directory, for example:

  programming/_index/
  programming/_index/all
  programming/_index/day
  programming/_index/month
  programming/_index/year

The above example is how you would define day, month, year, and all
indexes for the 'programming' category. If these files are found during
the publishing stage then they will be used to create the index of posts
for the journal.

Multiple remotes have been scrapped completely, in favour of a single
remote per journal. In my opinion this is simpler, and as far as I can
see there would be no reason for a single journal to have multiple
remotes.

Syntax highlighting for code blocks in pages, and posts is now handled
by the new 'gen-style' command. This will take the name of the style you
wish to apply to the blocks of code, and write the CSS of that style to
'stdout'. So if you would like to have syntax highlighting for your code
blocks simply do the following:

  $ jrnl gen-style solarized-light > _site/assets/style.css

Of course you would then include a link tag to the newly created
'style.css' file to your layout files where appropriate.

Other things of note, the `_meta.yml` file has been renamed to 'config'.
I find the name better encapsulates its purpose. And the 'command'
package has been renamed to 'cmd', just because I prefer 'cmd' over
'command'.

The 'usage' package has also been removed, this is only temporary.