Index published pages and posts in jrnl.hash file, and refactor
jrnl now keeps track of the published pages and posts in the jrnl.hash
file. This is a gob encoded map[string][]byte, where the []byte is the
SHA256 sum of the page's title, and body, and the string is the ID.

This hash is now used to only publish the pages/posts that have been
updated.