Fix edit bug, order categories, use paragraph for description
The 'jrnl edit' bug lead to posts/pages being overwritten completely, or
malformed when the 'jrnl edit' command was invoked. This was caused by
using the bufio.Reader in the unmarshalFrontMatter function. This
structure buffers the io.Reader in chunks of 4096. Because of this the
unread bytes aren't returned.

The categories are now ordered alphabetically.

Instead of using the first line on a post for the post's description,
the first paragraph is now used. If no paragraph can be found then it
falls back to the single new line.