Refactor and simplify
Removed a bunch of unnecessary dependencies. Publishing has now been
simplified. Previously we had indexes generated for the day, month and
year, this has been scrapped for simplicity. Now indexes are only
generated per category, and for the entire site via a special
_layouts/category-index and _layouts/index file respectively.

The commands 'title' and 'remote-set' have been removed, the 'config'
command is used instead for configuring the jrnl, of course you can also
choose to modify the toml file directly.

The cat command has been added for viewing the content of a page or a
post. The ls command comes with the -v switch for displaying has
information about the pages/posts and whether they have been modified.